[DHP-228] Moves disconnect modal logic to get triggered from DeviceDisconnectionCard component. Moves DisconnectionModal to its own file
