+++ src/applications/dhp-connected-devices/components/DeviceDisconnectionCard.jsx
@@ -1,30 +1,34 @@
-import React from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
+import { DisconnectModal } from './DisconnectModal';
 
-export const DeviceDisconnectionCard = ({ device, onClickHandler }) => {
+export const DeviceDisconnectionCard = ({ device }) => {
+  const [modalVisible, setModalVisible] = useState(false);
   return (
-    <div className="connect-device">
-      <h3 className="vads-u-margin-y--0">
-        {device.name}{' '}
-        <span className="connected-header-text"> - Connected</span>{' '}
-      </h3>
-      <p className="vads-u-margin-y--0">
-        <button
-          type="button"
-          onClick={onClickHandler}
-          data-testid={`${device.key}-disconnect-link`}
-          id={`${device.key}-disconnect-link`}
-          onKeyDown={onClickHandler}
-          className="usa-button-secondary"
-        >
-          Disconnect
-        </button>
-      </p>
-    </div>
+    <>
+      {modalVisible && <DisconnectModal deviceName={device.name} />}
+      <div className="connect-device">
+        <h3 className="vads-u-margin-y--0">
+          {device.name}{' '}
+          <span className="connected-header-text"> - Connected</span>{' '}
+        </h3>
+        <p className="vads-u-margin-y--0">
+          <button
+            type="button"
+            onClick={() => setModalVisible(true)}
+            data-testid={`${device.key}-disconnect-link`}
+            id={`${device.key}-disconnect-link`}
+            onKeyDown={() => setModalVisible(true)}
+            className="usa-button-secondary"
+          >
+            Disconnect
+          </button>
+        </p>
+      </div>
+    </>
   );
 };
 
 DeviceDisconnectionCard.propTypes = {
   device: PropTypes.object.isRequired,
-  onClickHandler: PropTypes.func.isRequired,
 };
